refactor #6 - 패키지 이름을 location 으로 변경합니다
