+++ src/main/java/com/depromeet/todo/domain/location/Location.java
@@ -1,0 +1,12 @@
+package com.depromeet.todo.domain.location;
+
+/**
+ * 위치
+ */
+public interface Location {
+    static Location of(Double latitude, Double longitude) {
+        return GcsLocation.of(latitude, longitude);
+    }
+
+    PcsLocation toPcsLocation();
+}
