+++ src/main/java/com/depromeet/todo/domain/location/PcsLocation.java
@@ -1,4 +1,4 @@
-package com.depromeet.todo.coordinate;
+package com.depromeet.todo.domain.location;
 
 import lombok.Value;
 
@@ -6,7 +6,12 @@ import lombok.Value;
  * 투영좌표계 (Projected Coordinate System, PCS)
  */
 @Value(staticConstructor = "of")
-public class PcsLocation {
+public class PcsLocation implements Location {
     public final Integer x;
     public final Integer y;
+
+    @Override
+    public PcsLocation toPcsLocation() {
+        return this;
+    }
 }
